+++ Application/Admin/Controller/GallaryController.class.php
@@ -7,6 +7,10 @@ class GallaryController extends Controller {
 	}
 
 	public function listAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$page = intval(I('get.page', '1'));
 		$amount = intval(I('get.pn', '10'));
 		$begin = intval(I('get.ps', ($page-1)*$amount));
@@ -34,6 +38,10 @@ class GallaryController extends Controller {
 	}
 
 	public function editAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$id = intval(I('get.id', '-1'));
 		$isnew = !($id >= 0);
 		$result = null;
@@ -50,6 +58,10 @@ class GallaryController extends Controller {
 	}
 
 	public function updateAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$Gallary = D('Gallary');
 		$result = $Gallary->create();
 		if ($result) $Gallary->save();
@@ -64,6 +76,10 @@ class GallaryController extends Controller {
 	}
 
 	public function createAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$Gallary = D('Gallary');
 		$result = $Gallary->create();
 		if ($result) $Gallary->add();
@@ -78,6 +94,10 @@ class GallaryController extends Controller {
 	}
 
 	public function deleteAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$id = intval(I('get.id', '-1'));
 		$Gallary = D('Gallary');
 		$result = $Gallary->delete($id);
@@ -92,6 +112,10 @@ class GallaryController extends Controller {
 	}
 
 	public function updateHandlerAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$Gallary = D('Gallary');
 		$result = $Gallary->create();
 		if ($result) $Gallary->save();
@@ -105,6 +129,10 @@ class GallaryController extends Controller {
 	}
 
 	public function adminPhotosAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$id = intval(I('get.id', '-1'));
 		$Gallary = D('Gallary');
 		$gallary = $Gallary->find($id);
@@ -119,6 +147,10 @@ class GallaryController extends Controller {
 	}
 
 	public function uploadPhotosAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$id = intval(I('get.id', '-1'));
 		$Gallary = D('Gallary');
 		$gallary = $Gallary->find($id);
@@ -130,6 +162,10 @@ class GallaryController extends Controller {
 	}
 
 	public function uploadPhotosHandlerAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$id = intval(I('get.id', '-1'));
 		$Gallary = D('Gallary');
 		$gallary = $Gallary->find($id);
@@ -179,6 +215,10 @@ class GallaryController extends Controller {
 	}
 
 	public function updatePhotoHandlerAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$Photo = D('Photo');
 		$result = $Photo->create();
 		if ($result) {
@@ -193,6 +233,10 @@ class GallaryController extends Controller {
 	}
 
 	public function deletePhotoHandlerAction() {
+		if (!APP_ADMIN) {
+			httpError(403, 'Not authorized');
+			return;
+		}
 		$Photo = D('Photo');
 		$thePhoto = $Photo->find(intval(I('post.id', '-1')));
 		$result = $Photo->where(array('id'=>intval(I('post.id', '-1'))))->delete();
